+++ src/main/java/tz/go/moh/him/thscp/mediator/elmis/MediatorMain.java
@@ -6,7 +6,7 @@ import akka.event.LoggingAdapter;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.openhim.mediator.engine.*;
-import tz.go.moh.him.thscp.mediator.elmis.DefaultOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.Orchestration.DefaultOrchestrator;
 
 import java.io.File;
 import java.io.IOException;
